fix hardwired SSL trust setting, refactor smtp config
