+++ src/main/java/uk/gov/hmcts/reform/sscs/SscsEvidenceShareApplication.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.reform.sscs;
 
-import java.util.Properties;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -8,8 +7,6 @@ import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.JavaMailSenderImpl;
 import org.springframework.retry.annotation.EnableRetry;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import uk.gov.hmcts.reform.authorisation.healthcheck.ServiceAuthHealthIndicator;
@@ -37,27 +34,6 @@ import uk.gov.hmcts.reform.sscs.ccd.config.CcdRequestDetails;
 @SuppressWarnings("HideUtilityClassConstructor") // Spring needs a constructor, its not a utility class
 public class SscsEvidenceShareApplication {
 
-    @Value("${appeal.email.host}")
-    private String emailHost;
-
-    @Value("${appeal.email.port}")
-    private int emailPort;
-
-    @Value("${appeal.email.smtp.tls.enabled}")
-    private String smtpTlsEnabled;
-
-    @Bean
-    public JavaMailSender javaMailSender() {
-        JavaMailSenderImpl javaMailSender = new JavaMailSenderImpl();
-        javaMailSender.setHost(emailHost);
-        javaMailSender.setPort(emailPort);
-        Properties properties = new Properties();
-        properties.setProperty("mail.transport.protocol","smtp");
-        properties.setProperty("mail.smtp.starttls.enable", smtpTlsEnabled);
-        properties.put("mail.smtp.ssl.trust","*");
-        javaMailSender.setJavaMailProperties(properties);
-        return javaMailSender;
-    }
 
     @Bean
     public CcdRequestDetails getRequestDetails(@Value("${core_case_data.jurisdictionId}") String coreCaseDataJurisdictionId,
