+++ src/main/java/uk/gov/hmcts/reform/sscs/config/SMTPConfig.java
@@ -1,0 +1,39 @@
+package uk.gov.hmcts.reform.sscs.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.mail.javamail.JavaMailSenderImpl;
+
+import java.util.Properties;
+
+@Configuration
+public class SMTPConfig {
+
+    @Value("${appeal.email.host}")
+    private String emailHost;
+
+    @Value("${appeal.email.port}")
+    private int emailPort;
+
+    @Value("${appeal.email.smtp.tls.enabled}")
+    private String smtpTlsEnabled;
+
+    @Value("${appeal.email.smtp.ssl.trust}")
+    private String smtpSslTrust;
+
+    @Bean
+    public JavaMailSender javaMailSender() {
+        JavaMailSenderImpl javaMailSender = new JavaMailSenderImpl();
+        javaMailSender.setHost(emailHost);
+        javaMailSender.setPort(emailPort);
+        Properties properties = new Properties();
+        properties.setProperty("mail.transport.protocol","smtp");
+        properties.setProperty("mail.smtp.starttls.enable", smtpTlsEnabled);
+        properties.put("mail.smtp.ssl.trust", smtpSslTrust);
+        javaMailSender.setJavaMailProperties(properties);
+        return javaMailSender;
+    }
+
+}
